Stricter compatibility with the getopt signature (returns an int, not a char)
gcc 4.7.2 in Ubuntu 12.10 detected an issue with assigning the result of getopt
to a char. When getopt returns -1 after exhausting the options, if it is
assigned to a char it sets all its bits to 1. When this result is then tested
for equality to the literal -1, the char was promoted to int, but the byte 0xff
turns into the signed integer 255. Aparently other compiling where assuming -1
is a char-type literal avoiding the conversions.